+++ src/eval/special.c
@@ -718,6 +718,7 @@ mino_val_t *eval_impl(mino_state_t *S, mino_val_t *form, mino_env_t *env, int ta
     case MINO_REGEX:
     case MINO_HOST_ARRAY:
     case MINO_MAP_ENTRY:
+    case MINO_TX_REF:
         return form;
     case MINO_SYMBOL:
         return eval_symbol(S, form, env);